+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     - "Expressions by Site"
     - "Randomizations by Site"
   - Add rake task to refresh reports on a scheduled basis
+  - Add sorting and searching to reports index
 - **Gem Changes**
   - Update to haml 5.1.1
   - Update to jbuilder 2.9